requirements.txt
django-environ==0.8.1
alabaster==0.7.13
arrow==1.2.3
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
Babel==2.11.0
binaryornot==0.4.4
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cookiecutter==2.1.1
cryptography==39.0.0
daphne==4.0.0
Deprecated==1.2.13
Django==4.1.6
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-model-utils==4.3.1
djangorestframework==3.14.0
docutils==0.19
ghp-import==2.1.0
gopay==1.3.0
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
incremental==22.10.0
Jinja2==3.1.2
jinja2-time==0.2.0
Markdown==3.3.7
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.4.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
Pillow==9.4.0
platformdirs==2.5.4
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7.1
PyYAML==6.0
pyyaml_env_tag==0.1
reportlab==3.6.12
requests==2.28.1
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib.applehelp==1.0.3
sqlparse==0.4.3
stripe==5.0.0
text-unidecode==1.3
tomli==2.0.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.13
watchdog==2.1.9
wrapt==1.14.1
zope.interface==5.5.2
whitenoise==5.3.0
boto3==1.16.56
django-storages==1.11.1
psycopg2-binary==2.9.2
